refactor(date-picker): use type format map internally
